requirements.txt
decord==0.6.0
einops==0.3.0
imageio==2.9.0
numpy==1.24.2
omegaconf==2.1.1
opencv_python
pandas==2.0.0
Pillow==9.5.0
PyYAML==6.0
setuptools==65.6.3
pytorch_lightning==1.9.3
tqdm==4.65.0
transformers==4.25.1
moviepy
av
xformers==0.0.22.post7
gradio
timm
scikit-learn 
open_clip_torch==2.22.0
kornia
tensorboard
fairscale
scikit-image==0.19.2
cupy-cuda12x
diffusers
transformers
einops
decord
tqdm
safetensors
omegaconf
opencv-python
pydantic
compel
easydict
rotary_embedding_torch
imageio[ffmpeg]
gradio
httpx[socks]
deepspeed==0.15.2
accelerate[deepspeed]
git+https://mirrors.tencent.com/github.com/microsoft/LoRA
huggingface_hub==0.25.0
lmdb